refactor: move fetch logic from table to manager
- simplifies table component which is now responsible only for presentation
